refactor(talk): keep control owner within lint limit
