+++ src/gateway/talk-client-pending-consult.ts
@@ -1,4 +1,14 @@
+import { BoundedSerialQueue } from "../shared/bounded-serial-queue.js";
 import { createDeferredCore } from "../shared/deferred.js";
+
+const REALTIME_CONTROL_MAX_PENDING = 8;
+
+export function createRealtimeControlQueue(): BoundedSerialQueue {
+  return new BoundedSerialQueue({
+    maxPendingCount: REALTIME_CONTROL_MAX_PENDING,
+    maxPendingWeight: REALTIME_CONTROL_MAX_PENDING,
+  });
+}
 
 export type PendingConsult<CapturedControl> = {
   controller: AbortController;
