+++ src/gateway/talk-realtime-relay-state.ts
@@ -15,7 +15,7 @@ import type { RealtimeVoiceSessionHarness } from "../talk/realtime-session-harne
 import type { RealtimeVoiceBridgeSession } from "../talk/session-runtime.js";
 import type { TalkEvent } from "../talk/talk-session-controller.js";
 import type { GatewayRequestContext } from "./server-methods/shared-types.js";
-import type { TalkAgentConsultAuthority } from "./talk-client-gateway-control.js";
+import type { TalkAgentConsultAuthority } from "./talk-agent-consult-authority.js";
 import type { RelayToolCallLedger } from "./talk-realtime-relay-tool-call-ledger.js";
 
 export const RELAY_SESSION_TTL_MS = 30 * 60 * 1000;
